Updated LinkedIn, GitHub, and resume urls to come from an environment variables
This is because those urls are subject to change. I'll be realistic, the GitHub and LinkedIn should
remain the same. However the resume url may change in the future. So when it does change, I can just
update the environment variable on Render, rather than having to make another commit to this repo.